all compiler warnings now fixed

+++ hail/src/main/scala/is/hail/annotations/Region.scala
@@ -85,19 +85,19 @@ object Region {
 
   def loadBit(byteOff: Long, bitOff: Long): Boolean = {
     val b = byteOff + (bitOff >> 3)
-    (loadByte(b) & (1 << (bitOff & 7))) != 0
+    (loadByte(b) & (1 << (bitOff & 7).toInt)) != 0
   }
 
   def setBit(byteOff: Long, bitOff: Long) {
     val b = byteOff + (bitOff >> 3)
     storeByte(b,
-      (loadByte(b) | (1 << (bitOff & 7))).toByte)
+      (loadByte(b) | (1 << (bitOff & 7).toInt)).toByte)
   }
 
   def clearBit(byteOff: Long, bitOff: Long) {
     val b = byteOff + (bitOff >> 3)
     storeByte(b,
-      (loadByte(b) & ~(1 << (bitOff & 7))).toByte)
+      (loadByte(b) & ~(1 << (bitOff & 7).toInt)).toByte)
   }
 
   def storeBit(byteOff: Long, bitOff: Long, b: Boolean) {
